% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arg_match}
\alias{arg_match}
\title{A better match.arg especially for checking function inputs}
\usage{
arg_match(arg)
}
\arguments{
\item{arg}{The name of the argument in the parent function}
}
\value{
the value of arg after verifying, or an error
}
\description{
Useful when specifying allowed values for a function argument in a vector in the formals. See examples.
}
\examples{
\dontrun{
my_fun <- function(my_arg = c("a", "b")) {
 my_arg <- arg_match(my_arg)
 return(paste("The value is", my_arg))
}

my_fun("a")
my_fun("b")
my_fun("c")
my_fun()
my_fun(NULL)
my_fun(5)
}
}
