\name{bcgroundwater-package}
\alias{bcgroundwater-package}
\alias{bcgroundwater}
\docType{package}
\title{
Work with B.C. groundwater level data
}
\description{
Set of functions for downloading, analysing, and visualising
groundwater level data from the BC Ministry of Environment groundwater
observation well network.
}
\details{
\tabular{ll}{
Package: \tab bcgroundwater\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-22\cr
License: \tab None Yet\cr
}

}
\author{
Andy Teucher <andy.teucher@gov.bc.ca>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
