% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWLdata.R
\name{get_gwl}
\alias{get_gwl}
\title{Retrive and format groundwater data from BC Government GWL site}
\usage{
get_gwl(well, which = "all", url = NULL, quiet = FALSE)
}
\arguments{
\item{well}{Character. The well number, accepts either \code{OW000} or
\code{000} format}

\item{which}{Character. Which data to retrieve \code{all} hourly data,
\code{recent} hourly data, or all \code{daily} averages.}

\item{url}{Character. Override the url location of the data.}

\item{quiet}{Logical. Suppress progress messages?}
}
\value{
A dataframe of the groundwater level observations
}
\description{
Go to <http://www.env.gov.bc.ca/wsd/data_searches/obswell/map/> to find your
well of interest.
}
\details{
Daily averages (\code{daily}) are calculated based on all Validated data.
}
\examples{
\dontrun{

all_309 <- get_gwl_data(well = 309)
recent_309 <- get_gwl_data(well = "OW309", which = "recent")
}

}
