% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlMonthlyPlot.R
\name{gwl_monthly_plot}
\alias{gwl_monthly_plot}
\title{Create a graph of historical monthly water level deviations}
\usage{
gwl_monthly_plot(dataframe, splines = TRUE, last12 = TRUE,
  save = FALSE, path = "./", opts = NULL)
}
\arguments{
\item{dataframe}{Dataframe containing 'Well_Num', 'Date', 'dev_med_GWL'}

\item{splines}{Logical: smooth the line using splines?}

\item{last12}{Logical: plot the last 12 monthly readings as points?}

\item{save}{Logical: save as a pdf?}

\item{path}{Path to folder in which to save if save=TRUE}

\item{opts}{Other options passed on to ggplot2}
}
\value{
A ggplot object.
}
\description{
Create a graph of historical median monthly water level deviations from
yearly average with 5th and 95th percentiles.
}
