% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlZypTest.R
\name{gwl_zyp_test}
\alias{gwl_zyp_test}
\title{Perform Mann-Kendall trend test on many wells}
\usage{
gwl_zyp_test(dataframe, wells = NULL, byID = "Well_Num",
  col = "mean_GWL", method = "both")
}
\arguments{
\item{dataframe}{Dataframe containing: an ID column (specified in \code{by})
and a column of values}

\item{wells}{Vector of well numbers to test. Default NULL does all in
dataframe}

\item{byID}{The name of the ID column}

\item{col}{The name of the column with the GWL values}

\item{method}{"both" (default), "yuepilon", or "zhang"}
}
\value{
A dataframe of results for all wells evaluated.
}
\description{
Uses the zyp package to calculate Mann-Kendall trend test on pre-whitened 
data (to remove autocorrelation) for many wells, using one or both of two 
pre-whitening methods, see zyp documentation.
}
