% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlMonthlyPlot.R
\name{gwlMonthlyPlot}
\alias{gwlMonthlyPlot}
\title{Create a graph of historical monthly water level deviations}
\usage{
gwlMonthlyPlot(dataframe, splines = TRUE, last12 = TRUE, save = FALSE,
  path = "./", opts = NULL)
}
\arguments{
\item{dataframe}{dataframe containing 'Well_Num', 'Date', 'dev_med_GWL'}

\item{splines}{logical: smooth the line using splines?}

\item{last12}{logical: plot the last 12 monthly readings as points?}

\item{save}{logical: save as a pdf?}

\item{path}{path to folder in which to save if save=TRUE}

\item{opts}{other options passed on to ggplot2}
}
\value{
A ggplot object
}
\description{
Create a graph of historical median monthly water level deviations from yearly 
average with 5th and 95th percentiles
}
\examples{
\dontrun{

}

}
