\name{knitr_latex_char}
\alias{knitr_latex_char}
\title{Replace characters with LaTeX compatible characters/code}
\usage{
knitr_latex_char(x)
}
\arguments{
  \item{x}{String containing invalid LaTeX characters}
}
\value{
String with LaTeX compatbible characters
}
\description{
Replaces or escapes characters (\,#,$,%,&,_,^,~) that can't
be used directly in LaTeX. Does not deal with { and }
}
\examples{
\dontrun{

}
}

