% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWellTS.R
\name{make_well_ts}
\alias{make_well_ts}
\alias{makeWellTS}
\title{Get full time series with missing values interpolated}
\usage{
make_well_ts(df, trim = TRUE, head = 0.1, tail = 0.9, n_consec = 4)
}
\arguments{
\item{df}{A monthly dataframe created by `monthlyValues`. Must minimally 
include fields `EMS_ID`, `Well_Num` `Date`, 'med_GWL`, `nReadings`}

\item{trim}{Should leading and trailing missing values in be removed? See [trim_cons_runs()].}

\item{head}{(Proportion between 0 and 1, default = 0.1) The beginning portion of 
the vector in which to look for consecutive missing values}

\item{tail}{(Proportion between 0 and 1, default = 0.9) The end portion of 
the vector in which to look for consecutive missing values}

\item{n_consec}{The number of consecutive values which constitutes a 'long' run (default = 4)}
}
\value{
A full monthly time series with interpolated missing values, retaining
 all of the columns in the original data frame.
}
\description{
Takes a dataframe with monthly values and creates a full time series, 
interpolating missing values.
}
