% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeWellTS}
\alias{makeWellTS}
\title{Get full time series with missing values interpolated}
\usage{
makeWellTS(df)
}
\arguments{
\item{df}{A monthly dataframe created by `monthlyValues`. Must minimally include
fields `EMS_ID`, `Well_Num` `Date`, 'med_GWL`}
}
\value{
A full monthly time series with interpolated missing values,
        retaining all of the columns in the original data frame.
}
\description{
Takes a dataframe with monthly values and creates a full time series, interpolating
missing values.
}
\examples{
\dontrun{

}
}

