% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPointWithInset.R
\name{plot_point_with_inset}
\alias{plot_point_with_inset}
\title{Plot a point on a map with an overview inset}
\usage{
plot_point_with_inset(long, lat, pointColour, bigMap = NULL,
  overviewMap = NULL, overviewExtent = c(-139, 48, -114, 60),
  opts = NULL)
}
\arguments{
\item{long}{Longitude (dd) of point}

\item{lat}{Latitude (dd) of point}

\item{pointColour}{The colour of the point}

\item{bigMap}{ggmap of the zoomed in area that will comprise the large map}

\item{overviewMap}{ggMap of large area that will be in the inset.  If not
provided, will be generated using overviewExtent}

\item{overviewExtent}{The extent (left, botton, right, top) of the large
area that will be in the inset. Default is British Columbia}

\item{opts}{Additional options to pass to ggplot2}
}
\value{
A ggmap object.
}
\description{
Plot a point on a map with an inset in the top right corner showing an 
overview of a larger area.
}
