\name{readGWLdata}
\alias{readGWLdata}
\title{Read in groundwater data from file downloaded from GWL tool}
\usage{
readGWLdata(emsID, wellNum, dir)
}
\arguments{
  \item{emsID}{The EMS ID of the well}

  \item{wellNum}{The Well Number of the well}

  \item{dir}{The directory the file is stored in}
}
\value{
A dataframe of the groundwater level observations
}
\description{
Read in groundwater data from file downloaded from GWL tool
}
\examples{
\dontrun{

}
}

