% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readGWLdata2.R
\name{readGWLdata2}
\alias{readGWLdata2}
\title{Read in groundwater data from files downloaded from GWL tool}
\usage{
readGWLdata2(historical, datafile, stnName = "Not Available", emsID = "NA")
}
\arguments{
\item{historical}{The path to the csv with historical data}

\item{datafile}{The path to the csv with current data}

\item{stnName}{The station name to use as it is no longer in data}

\item{emsID}{The EMS ID of the well retained for compatibility}
}
\value{
A dataframe of the groundwater level observations
}
\description{
Read in groundwater data from files downloaded from GWL tool
}
\examples{
\dontrun{

}
}
\author{
Jay R Brown, \email{jay@systemicresult.com}
}

