% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimConsRuns.R
\name{trim_cons_runs}
\alias{trim_cons_runs}
\alias{trimConsRuns}
\title{Find consecuctive runs of a value in extremes of a vector}
\usage{
trim_cons_runs(x, val, head = 0.1, tail = 0.9, n_consec = 4)
}
\arguments{
\item{x}{The vector}

\item{val}{The value for which to find consecutive runs}

\item{head}{(Proportion between 0 and 1, default = 0.1) The beginning portion of 
the vector in which to look for consecutive missing values}

\item{tail}{(Proportion between 0 and 1, default = 0.9) The end portion of 
the vector in which to look for consecutive missing values}

\item{n_consec}{The number of consecutive values which constitutes a 'long' run (default = 4)}
}
\value{
A list of start and end points of a vector to remove heavily weighted
        heads and tails containing continuous runs of a value (likely denoting missing)
}
\description{
Finds the start and end points of consecutive runs of a value in the 
beginning and end of a vector.
}
\examples{
\dontrun{
 x <- rbinom(100, size=1, prob=0.5)
 trunc <- trimConsRuns(x, 1, n_consec=4)
 trunc_x <- x[trunc$start:trunc$end]
}
}
