% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgroundwater.R
\docType{package}
\name{bcgroundwater}
\alias{bcgroundwater}
\alias{bcgroundwater-package}
\title{bcgroundwater: Analyse B.C. groundwater level data.}
\description{
Set of functions for downloading, analysing, and visualising
groundwater level data from the BC Ministry of Environment groundwater
observation well network.
}
\author{
Andy Teucher \email{andy.teucher@gov.bc.ca}
}

