% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWLdata.R
\name{get_gwl}
\alias{get_gwl}
\title{Retrive and format groundwater data from BC Government GWL site}
\usage{
get_gwl(well, which = "all", url = NULL, quiet = FALSE)
}
\arguments{
\item{well}{Character. The well number, accepts either \code{OW000} or 
\code{000} format}

\item{which}{Character. Which data to retrieve \code{all} hourly data, 
\code{recent} hourly data, or all \code{daily} averages.}

\item{url}{Character. Override the url location of the data.}

\item{quiet}{Logical. Suppress progress messages?}
}
\value{
A dataframe of the groundwater level observations
}
\description{
Go to <http://www.env.gov.bc.ca/wsd/data_searches/obswell/map/> to find your 
well of interest.
}
\details{
Note that well water levels are measured in meters below the ground. Thus 
higher values represent deeper levels. Therefore 
\code{Historical_Daily_Minimum} values will be greater than 
\code{Historical_Daily_Maximum} values, as they represent a lower water 
level.

Daily averages (\code{which = 'daily'}) are pre-calculated. Note that they 
only cover days marked as "Validated" in the full hourly dataset.
}
\examples{
\dontrun{

all_309 <- get_gwl(well = 309)
recent_309 <- get_gwl(well = "OW309", which = "recent")
daily_avg_309 <- get_gwl(well = "OW309", which = "daily")
}

}
