% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWLdata.R
\name{get_gwl}
\alias{get_gwl}
\title{Retrieve and format groundwater data from B.C. Data Catalogue}
\usage{
get_gwl(wells, which = c("all", "recent", "daily"), url = NULL,
  quiet = FALSE)
}
\arguments{
\item{wells}{Character vector. The well number(s), accepts either 
\code{OW000} or \code{000} format. Note that format OW000 requires three
digit numbers, e.g. OW309, OW008, etc.}

\item{which}{Character. Which data to retrieve \code{all} hourly data, 
\code{recent} hourly data, or all \code{daily} averages}

\item{url}{Character. Override the url location of the data}

\item{quiet}{Logical. Suppress progress messages?}
}
\value{
A dataframe of the groundwater level observations.
}
\description{
Go to <https://www2.gov.bc.ca/gov/content?id=2D3DB9BA78DE4377AB4AE2DCEE1B409B> to find your 
well(s) of interest.
}
\details{
Note that well water levels are measured in meters below the ground. Thus 
higher values represent deeper levels. Therefore 
\code{Historical_Daily_Minimum} values will be greater than 
\code{Historical_Daily_Maximum} values, as they represent a lower water 
level.

Daily averages (\code{which = 'daily'}) are pre-calculated. Note that they 
only cover days marked as "Validated" in the full hourly dataset.
}
\examples{
\dontrun{

all_309 <- get_gwl(wells = 309)
recent_309 <- get_gwl(wells = "OW309", which = "recent")
daily_avg_309 <- get_gwl(wells = "OW309", which = "daily")

all_multi < get_gwl(wells = c(309, 89))
recent_multi <- get_gwl(wells = c("OW309", "OW089"), which = "recent")
daily_avg_multi <- get_gwl(wells = c("309", "89"), which = "daily")
}

}
