% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlAreaPlot.R
\name{gwl_area_plot}
\alias{gwl_area_plot}
\alias{gwlAreaPlot}
\title{Creates an area plot of groundwater levels with trend line}
\usage{
gwl_area_plot(dataframe, trend, intercept, state, sig,
  showInterpolated = FALSE, save = FALSE, path = "./",
  mkperiod = "monthly", opts = NULL)
}
\arguments{
\item{dataframe}{A dataframe with well level monthly time series and the 
following columns: Date, med_GWL, nReadings}

\item{trend}{(Numeric) Trend in m/month}

\item{intercept}{(Numeric) Intercept in m}

\item{state}{Trend classification: \code{"stable"}, or another description 
of the magnitude and direction of the trend. This description will
be displayed on the graph in Title Case. 
If \code{"stable"}, no trend value or statistical significance 
will be displayed on the plot, otherwise this information will be 
displayed alongside the \code{state} information.}

\item{sig}{(Numeric) Significance of trend test}

\item{showInterpolated}{(Logical) Show the points where missing values in the 
time series were interpolated}

\item{save}{Save the graph to file?}

\item{path}{Where to save the graph if \code{save = TRUE}}

\item{mkperiod}{The period (\code{"monthly"} or \code{"annual"}) 
the Mann-Kendall test was performed on}

\item{opts}{Other options to pass to ggplot2}
}
\value{
A ggplot2 object.
}
\description{
An area plot (hydrograph) of groundwater levels, with trend line of a 
given slope and intercept, optionally with interpolated values shown.
}
