% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwlAreaPlot.R
\name{gwlAreaPlot}
\alias{gwlAreaPlot}
\title{Creates an area plot of groundwater levels with trend line}
\usage{
gwlAreaPlot(dataframe, trend, intercept, state, sig, showInterpolated = FALSE,
  save = FALSE, path = "./", mkperiod = "monthly", opts = NULL)
}
\arguments{
\item{dataframe}{a dataframe with well level monthly time series and the 
following columns: Date, med_GWL, nReadings}

\item{trend}{(numeric) trend in m/month}

\item{intercept}{(numeric) intercept in m}

\item{state}{Trend classification (stable, declining, increasing)}

\item{sig}{(numeric) significance of trend test}

\item{showInterpolated}{(logical) show the points where missing values in the 
time series were interpolated}

\item{save}{Save the graph to file?}

\item{path}{Where to save the graph if \code{save = TRUE}}

\item{mkperiod}{the period (monthly or annual) the Mann-Kendall test was 
performed on}

\item{opts}{other options to pass to ggplot2}
}
\value{
A ggplot2 object
}
\description{
An area plot (hydrograph) of groundwater levels, with trend line of a 
given slope and intercept, optionally with interpolated values shown.
}
