% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWellTS.R
\name{makeWellTS}
\alias{makeWellTS}
\title{Get full time series with missing values interpolated}
\usage{
makeWellTS(df, trim = TRUE, ...)
}
\arguments{
\item{df}{A monthly dataframe created by `monthlyValues`. Must minimally 
include fields `EMS_ID`, `Well_Num` `Date`, 'med_GWL`, `nReadings`}

\item{trim}{Should leading and trailing missing values in be removed? See [trimConsRuns()].}

\item{...}{Parameters passed on to [trimConsRuns()] (`head`, `tail`, `n_consec`)}
}
\value{
A full monthly time series with interpolated missing values, retaining
 all of the columns in the original data frame.
}
\description{
Takes a dataframe with monthly values and creates a full time series, 
interpolating missing values.
}
\examples{
\dontrun{

}

}
