% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readGWLdata}
\alias{readGWLdata}
\title{Read in groundwater data from file downloaded from GWL tool}
\usage{
readGWLdata(path, emsID)
}
\arguments{
\item{path}{The path to the csv}

\item{emsID}{The EMS ID of the well}
}
\value{
A dataframe of the groundwater level observations
}
\description{
Read in groundwater data from file downloaded from GWL tool
}
\examples{
\dontrun{

}
}

