% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWLdata.R
\name{readGWLdata}
\alias{readGWLdata}
\title{Read in groundwater data from file downloaded from GWL tool}
\usage{
readGWLdata(path, emsID = NULL)
}
\arguments{
\item{path}{The path to the csv}

\item{emsID}{The EMS ID of the well}
}
\value{
A dataframe of the groundwater level observations
}
\description{
Read in groundwater data from file downloaded from GWL tool
}
\examples{
\dontrun{

}
}

