% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_dd.R
\name{utm_dd}
\alias{utm_dd}
\title{Convert zone+utm pairs to lat/long}
\usage{
utm_dd(zone = NULL, easting = NULL, northing = NULL,
  datum = "NAD83", data = NULL, key = NULL)
}
\arguments{
\item{zone}{Integer, or column name in 'data'}

\item{easting}{Integer, or column name in 'data'}

\item{northing}{Integer, or column name in 'data'}

\item{datum}{String or column name in 'data'. Default 'NAD83'}

\item{data}{(Optional) A data frame with zone + utms}

\item{key}{Name of column in 'data' that contains a unique identifier for each row}
}
\value{
Either a vector of length 2 with longitude and latitude
       respectively, or a dataframe with the 'key' column and longitude
       and latitude.
}
\description{
Convert zone+utm pairs to lat/long. Can take either a single zone + utm,
or a data frame with many.  Allows for datasets with different zones and 
datum for each utm.
}
